+++ src/GafferScene/AttributeVisualiser.cpp
@@ -273,6 +273,30 @@ IECore::ConstCompoundObjectPtr AttributeVisualiser::computeProcessedAttributes( 
 			case Color3fDataTypeId :
 				color = static_cast<const Color3fData *>( attribute )->readable();
 				break;
+			case V2iDataTypeId : {
+				const auto &v = static_cast<const V2iData *>( attribute )->readable();
+				color = Color3f( v.x, v.y, 0 );
+				break;
+			}
+			case V2fDataTypeId : {
+				const auto &v = static_cast<const V2fData *>( attribute )->readable();
+				color = Color3f( v.x, v.y, 0 );
+				break;
+			}
+			case V2dDataTypeId : {
+				const auto &v = static_cast<const V2dData *>( attribute )->readable();
+				color = Color3f( v.x, v.y, 0 );
+				break;
+			}
+			case V3iDataTypeId :
+				color = static_cast<const V3iData *>( attribute )->readable();
+				break;
+			case V3fDataTypeId :
+				color = static_cast<const V3fData *>( attribute )->readable();
+				break;
+			case V3dDataTypeId :
+				color = static_cast<const V3dData *>( attribute )->readable();
+				break;
 			default :
 				throw IECore::Exception( boost::str(
 					boost::format( "Unsupported attribute data type \"%s\"" ) % attribute->typeName()